Fix for #97: handle correctly flow rotations
The flow ought to be transformed with respect to the coordinates where
it lands.

Also, small angle approximation is not used here. This adds a few more
computations but should be manageable in terms of execution time.
Moreover, this opens up the door for general linear transforms only by
tweaking the coefficients.

I measured about +30% execution time increase with this version. That
seems needed, however, if we look at how off the transformed flow is
with large displacements (eg. FlyingChairs dataset).